feat!: update to ver 2.0.0 🎉
