+++ .gitignore
@@ -8,14 +8,9 @@ pnpm-debug.log*
 lerna-debug.log*
 
 node_modules
+dist
 dist-ssr
-build
 *.local
-
-# dist ignore unnecessary files
-dist/index.html
-dist/pvue.js
-dist/vite.svg
 
 # Editor directories and files
 .vscode/*
@@ -27,3 +22,4 @@ dist/vite.svg
 *.njsproj
 *.sln
 *.sw?
+release
