+++ package.json
@@ -1,20 +1,33 @@
 {
-  "name": "vite-project",
+  "name": "vue-mini-player-next",
   "private": true,
   "version": "0.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "tsc && vite build",
-    "preview": "vite preview"
+    "preview": "vite preview",
+    "rollup": "rollup -c"
   },
   "devDependencies": {
-    "@types/lodash": "^4.14.192",
-    "@types/node": "^18.11.9",
-    "typescript": "^4.6.4",
-    "vite": "^3.2.3"
+    "@rollup/plugin-commonjs": "^25.0.3",
+    "@rollup/plugin-node-resolve": "^15.1.0",
+    "@rollup/plugin-typescript": "^11.1.2",
+    "@types/lodash": "^4.14.195",
+    "postcss": "^8.4.26",
+    "postcss-url": "^10.1.3",
+    "rollup": "^3.26.3",
+    "rollup-plugin-ignore": "^1.0.10",
+    "rollup-plugin-ignore-import": "^1.3.2",
+    "rollup-plugin-postcss": "^4.0.2",
+    "rollup-plugin-string": "^3.0.0",
+    "rollup-plugin-terser": "^7.0.2",
+    "typescript": "^5.1.6",
+    "vite": "^4.4.0",
+    "vite-plugin-string": "^1.2.1"
   },
   "dependencies": {
-    "lodash": "^4.17.21"
+    "lodash": "^4.17.21",
+    "petite-vue": "^0.4.1"
   }
 }
