+++ src/core/MusicPlayerCore/core.d.ts
@@ -1,0 +1,13 @@
+
+export interface SingleSongBriefInfo {
+  name: string,
+  id: string,
+  src: string,
+  author?: string[],
+  album?: string,
+  img?: string
+}
+
+export interface SingleSongProps extends SingleSongBriefInfo {
+
+}
