+++ src/core/MusicPlayerCore/utils.ts
@@ -1,0 +1,7 @@
+export function throwError(info: string) {
+  throw new Error('[PlayerCore error]: ' + info)
+}
+
+export function logWarn(info: string) {
+  console.warn('[PlayerCore warn]: ' + info)
+}
