--- src/core/core.d.ts
@@ -1,13 +1,0 @@
-
-export interface SingleSongBriefInfo {
-  name: string,
-  id: string,
-  src: string,
-  author?: string[],
-  album?: string,
-  img?: string
-}
-
-export interface SingleSongProps extends SingleSongBriefInfo {
-
-}
