+++ src/css/index.css
@@ -1,10 +1,3 @@
-.player_module {
-  background-image: linear-gradient(90deg, rgba(99, 99, 99, 0.1), rgba(99, 99, 99, 0.1) 1px, transparent 0, transparent), linear-gradient(90deg, hsla(0, 0%, 100%, 0.1) -20%, transparent 15%), linear-gradient(0deg, hsla(0, 0%, 100%, 0.1), transparent 25%);
-  background-color: rgba(0, 0, 0, 0.6);
-  border-top: 1px solid hsla(0, 0%, 100%, 0.49411765);
-  border-radius: 10px;
-  box-shadow: 0 0 10px 5px hsla(0, 0%, 100%, 0.15);
-}
 #CorePlayer {
   position: fixed;
   width: 0;
@@ -12,11 +5,17 @@
   right: 0;
   bottom: 0;
 }
+#CorePlayer .player_module {
+  background-image: linear-gradient(90deg, rgba(99, 99, 99, 0.1), rgba(99, 99, 99, 0.1) 1px, transparent 0, transparent), linear-gradient(90deg, hsla(0, 0%, 100%, 0.1) -20%, transparent 15%), linear-gradient(0deg, hsla(0, 0%, 100%, 0.1), transparent 25%);
+  background-color: rgba(0, 0, 0, 0.6);
+  border-top: 1px solid hsla(0, 0%, 100%, 0.49411765);
+  border-radius: 10px;
+  box-shadow: 0 0 10px 5px hsla(0, 0%, 100%, 0.15);
+}
 #CoreWrapper {
   display: flex !important;
   font-family: 'Microsoft Yahei', 'Times New Roman', Times, serif;
   bottom: 0;
-  left: 0;
   position: fixed;
   width: 400px;
   height: 120px;
