+++ src/css/index.less
@@ -1,24 +1,23 @@
-.player_module {
-  background-image: linear-gradient(90deg, rgba(99, 99, 99, .1), rgba(99, 99, 99, .1) 1px, transparent 0, transparent), linear-gradient(90deg, hsla(0, 0%, 100%, .1) -20%, transparent 15%), linear-gradient(0deg, hsla(0, 0%, 100%, .1), transparent 25%);
-  background-color: rgba(0, 0, 0, .6);
-  border-top: 1px solid hsla(0, 0%, 100%, .4941176471);
-  border-radius: 10px;
-  box-shadow: 0 0 10px 5px hsla(0, 0%, 100%, .15);
-}
-
 #CorePlayer {
   position: fixed;
   width: 0;
   height: 0;
   right: 0;
   bottom: 0;
+
+  .player_module {
+    background-image: linear-gradient(90deg, rgba(99, 99, 99, .1), rgba(99, 99, 99, .1) 1px, transparent 0, transparent), linear-gradient(90deg, hsla(0, 0%, 100%, .1) -20%, transparent 15%), linear-gradient(0deg, hsla(0, 0%, 100%, .1), transparent 25%);
+    background-color: rgba(0, 0, 0, .6);
+    border-top: 1px solid hsla(0, 0%, 100%, .4941176471);
+    border-radius: 10px;
+    box-shadow: 0 0 10px 5px hsla(0, 0%, 100%, .15);
+  }
 }
 
 #CoreWrapper {
   display: flex !important;
   font-family: 'Microsoft Yahei', 'Times New Roman', Times, serif;
   bottom: 0;
-  left: 0;
   position: fixed;
   width: 400px;
   height: 120px;
