+++ src/pack/WithoutPvue.ts
@@ -1,0 +1,4 @@
+/** pack without petite-vue */
+import VueMiniPlayerCore from '../core/VueMiniPlayerCore'
+
+export default VueMiniPlayerCore
